requirements.txt
requests
pyyaml
pytz
akshare
pandas
